+++ packages/@core/type-utils/src/index.ts
@@ -1,7 +1,12 @@
-export * from './fn';
+/**
+ * 平台中立的类型工具出口。
+ *
+ * 依赖 DOM 的表单元素类型在 `@skyroc/type-utils/web`。
+ */
+export type * from './fn';
 
-export type * from './form';
+export type * from './object';
 
-export * from './utility-types';
+export type * from './path';
 
-export * from './utils';
+export type * from './primitive';